Fix Get-Process, Start-Sleep and some tests
