+++ appveyor.yml
@@ -21,4 +21,3 @@ after_test:
 
 on_finish: 
   - ps: Invoke-AppveyorFinish
-  - ps: $blockRdp = $true; iex ((new-object net.webclient).DownloadString('https://raw.githubusercontent.com/appveyor/ci/master/scripts/enable-rdp.ps1'))
