+++ src/Microsoft.PowerShell.Commands.Utility/commands/utility/StartSleepCommand.cs
@@ -85,11 +85,7 @@ namespace Microsoft.PowerShell.Commands
             }
             if (_waitHandle != null)
             {
-#if CORECLR //TODO:CORECLR bool WaitOne(int millisecondsTimeout,bool exitContext) is not available on CLR yet
-                _waitHandle.WaitOne(new TimeSpan(0, 0, 0, 0, milliSecondsToSleep));
-#else
-                _waitHandle.WaitOne(new TimeSpan(0, 0, 0, 0, milliSecondsToSleep), true);
-#endif
+                _waitHandle.WaitOne(milliSecondsToSleep, true);
             }
         }
