+++ test/powershell/Modules/Microsoft.PowerShell.Utility/Start-Sleep.Tests.ps1
@@ -1,18 +1,18 @@
 Describe "Start-Sleep DRT Unit Tests" -Tags "CI" {
-    It "Should be works properly when sleeping with Second" {
+    It "Should work properly when sleeping with Second" {
         $dtStart = [DateTime]::Now
         Start-Sleep -Seconds 1
         $dtEnd = [DateTime]::Now
         $milliseconds = (New-TimeSpan -Start $dtStart -End $dtEnd).TotalMilliseconds
-        $milliseconds | Should BeGreaterThan 1000
+        $milliseconds | Should BeGreaterThan 999
     }
 
-    It "Should be works properly when sleeping with Milliseconds" {
+    It "Should work properly when sleeping with Milliseconds" {
         $dtStart = [DateTime]::Now
         Start-Sleep -Milliseconds 1000
         $dtEnd = [DateTime]::Now
         $milliseconds = (New-TimeSpan -Start $dtStart -End $dtEnd).TotalMilliseconds
-        $milliseconds -ge 1000 | Should Be $true
+        $milliseconds | Should BeGreaterThan 999
     }
 
 }
